Add a new cpuidle-smt-performance test
This creates a pair of threads thread0 and thread1, each of which can
be pinned to a CPU. thread0 CPU can be provided using the option
--pcpu or -p and the thread1 CPU can be provided using the option
--ccpu or -c.

Each thread runs for a specified duration (default 100us) while the
other one sleeps. When one of the threads finishes running for the
said duration, it wakes up the other thread and goes to sleep. This
continues until the timeout (default 5 seconds).

Each thread makes a note of the time it took to wakeup the other
thread, as well as the number of operations (each thread performs
fibonacci computation during its runtime duration) that it was perform
while it was running.

How to use this benchmark:
~~~~~~~~~~~~~~~~~~~~~~~~~~~~

Pin thread0 and thread1 to the CPUs of the same core. Provide a
runtime corresponding to the cpuidle residency of the desired state,
so that when a thread is idling, the CPU goes to that idle state.

The greater the duration, the deeper the idle state, and the higher
the chances of boosting the single threaded performance. But there is
a penalty in terms of wakeup latency deeper the state.

$ ./cpuidle-smt-performance -p 0 -c 1 -r 10
Thread0[PID 27230] affined to CPUs: 0,
Thread1[PID 27231] affined to CPUs: 1,
===============================================
                  Summary
===============================================
fib_iteration duration to 10 us
Total number of operations = 156.178 Million ops/seconds
Average wakeup latency = 3.942 us
===============================================

$ ./cpuidle-smt-performance -p 0 -c 1 -r 50
Thread0[PID 27264] affined to CPUs: 0,
Thread1[PID 27265] affined to CPUs: 1,
===============================================
                  Summary
===============================================
fib_iteration duration to 50 us
Total number of operations = 196.012 Million ops/seconds
Average wakeup latency = 4.017 us
===============================================

$./cpuidle-smt-performance -p 0 -c 1 -r 100
Thread0[PID 27274] affined to CPUs: 0,
Thread1[PID 27275] affined to CPUs: 1,
===============================================
                  Summary
===============================================
fib_iteration duration to 100 us
Total number of operations = 199.906 Million ops/seconds
Average wakeup latency = 4.799 us
===============================================

$ ./cpuidle-smt-performance -p 0 -c 1 -r 500
Thread0[PID 27308] affined to CPUs: 0,
Thread1[PID 27309] affined to CPUs: 1,
===============================================
                  Summary
===============================================
fib_iteration duration to 500 us
Total number of operations = 204.066 Million ops/seconds
Average wakeup latency = 5.728 us
===============================================

$ ./cpuidle-smt-performance -p 0 -c 1 -r 1000
Thread0[PID 27351] affined to CPUs: 0,
Thread1[PID 27352] affined to CPUs: 1,
===============================================
                  Summary
===============================================
fib_iteration duration to 1000 us
Total number of operations = 205.599 Million ops/seconds
Average wakeup latency = 7.330 us

Signed-off-by: Gautham R. Shenoy <ego@linux.vnet.ibm.com>